LRCI-428 Extract common test suite logic from PortalAcceptancePullRequestJob into PortalAcceptanceTestSuiteJob
