+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/PortalAcceptancePullRequestJob.java
@@ -16,7 +16,6 @@ package com.liferay.jenkins.results.parser;
 
 import java.io.File;
 
-import java.util.Properties;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -24,7 +23,7 @@ import java.util.TreeSet;
  * @author Michael Hashimoto
  */
 public class PortalAcceptancePullRequestJob
-	extends PortalGitRepositoryJob implements TestSuiteJob {
+	extends PortalAcceptanceTestSuiteJob {
 
 	public PortalAcceptancePullRequestJob(String jobName) {
 		this(jobName, "default");
@@ -33,24 +32,12 @@ public class PortalAcceptancePullRequestJob
 	public PortalAcceptancePullRequestJob(
 		String jobName, String testSuiteName) {
 
-		super(jobName);
-
-		_testSuiteName = testSuiteName;
+		super(jobName, testSuiteName);
 	}
 
 	@Override
 	public Set<String> getBatchNames() {
-		Properties jobProperties = getJobProperties();
-
-		String testBatchNames = JenkinsResultsParserUtil.getProperty(
-			jobProperties, "test.batch.names[" + _testSuiteName + "]");
-
-		if (testBatchNames == null) {
-			testBatchNames = JenkinsResultsParserUtil.getProperty(
-				jobProperties, "test.batch.names");
-		}
-
-		Set<String> testBatchNamesSet = getSetFromString(testBatchNames);
+		Set<String> testBatchNamesSet = super.getBatchNames();
 
 		if (_isRelevantTestSuite() && _isPortalWebOnly()) {
 			String[] portalWebOnlyBatchNameMarkers = {
@@ -77,53 +64,6 @@ public class PortalAcceptancePullRequestJob
 		return testBatchNamesSet;
 	}
 
-	@Override
-	public Set<String> getDistTypes() {
-		Properties jobProperties = getJobProperties();
-
-		String testBatchDistAppServers = JenkinsResultsParserUtil.getProperty(
-			jobProperties,
-			"test.batch.dist.app.servers[" + _testSuiteName + "]");
-
-		if (testBatchDistAppServers == null) {
-			testBatchDistAppServers = JenkinsResultsParserUtil.getProperty(
-				jobProperties, "test.batch.dist.app.servers");
-		}
-
-		return getSetFromString(testBatchDistAppServers);
-	}
-
-	@Override
-	public String getPoshiQuery(String testBatchName) {
-		String[] propertyNames = {
-			JenkinsResultsParserUtil.combine(
-				"test.batch.run.property.query[", testBatchName, "][",
-				_testSuiteName, "]"),
-			JenkinsResultsParserUtil.combine(
-				"test.batch.run.property.query[", testBatchName, "]")
-		};
-
-		Properties jobProperties = getJobProperties();
-
-		for (String propertyName : propertyNames) {
-			if (jobProperties.containsKey(propertyName)) {
-				String propertyValue = JenkinsResultsParserUtil.getProperty(
-					jobProperties, propertyName);
-
-				if ((propertyValue != null) && !propertyValue.isEmpty()) {
-					return propertyValue;
-				}
-			}
-		}
-
-		return null;
-	}
-
-	@Override
-	public String getTestSuiteName() {
-		return _testSuiteName;
-	}
-
 	private boolean _isPortalWebOnly() {
 		GitWorkingDirectory gitWorkingDirectory = getGitWorkingDirectory();
 
@@ -142,9 +82,9 @@ public class PortalAcceptancePullRequestJob
 	}
 
 	private boolean _isRelevantTestSuite() {
-		return _testSuiteName.equals("relevant");
-	}
+		String testSuiteName = getTestSuiteName();
 
-	private final String _testSuiteName;
+		return testSuiteName.equals("relevant");
+	}
 
 }
